requirements.txt
aiogram>=3.0
requests
openpyxl
yadisk
python-dotenv
PyYAML